feat: 動的テキストサイズ調整とポート管理機能を追加
- ActionButtonに動的フォントサイズ調整機能を実装
  - テキスト長に応じた自動フォントサイズ計算
  - プラットフォームネイティブのtext-overflow: ellipsis対応
  - CSS Gridレイアウトによる改良されたボタン構造
  - DPI対応とレスポンシブデザインの強化

- launch.ps1に包括的なポート管理機能を追加
  - 動的ポート検出と自動割り当て
  - インテリジェントプロセス管理（開発関連プロセスのみ終了）
  - 環境変数による動的ポート設定
  - カスタマイズ可能なパラメータ（-StartPort, -MaxPortAttempts）

- vite.config.tsを環境変数対応に更新
  - VITE_PORT環境変数による動的ポート設定
  - strictPortをfalseに変更してフォールバック対応

- アイコン処理システムの実装
  - Rustバックエンドでのアイコン抽出とキャッシュ機能
  - 複数アイコン形式のサポート（Emoji, File, URL, Base64）
  - フロントエンドでの非同期アイコン処理

- テストの更新と改善
  - 新機能に対応したテストケースの追加
  - モックデータの更新
  - 25個のテストが全て成功

これらの改善により、文字のはみ出し問題が解決され、
ポート競合も自動的に回避されるようになりました。

+++ src-tauri/src/modules/hotkey.rs
@@ -559,7 +559,7 @@ mod tests {
         
         // Test basic hotkey parsing
         let parsed = service.parse_hotkey_string("Ctrl+Alt+F1").unwrap();
-        assert_eq!(parsed.modifiers, MOD_CONTROL | MOD_ALT);
+        assert_eq!(parsed.modifiers, (MOD_CONTROL | MOD_ALT) as u32);
         assert_eq!(parsed.vk_code, 0x70); // F1
         
         // Test single key
@@ -569,7 +569,7 @@ mod tests {
         
         // Test letter key
         let parsed = service.parse_hotkey_string("Ctrl+A").unwrap();
-        assert_eq!(parsed.modifiers, MOD_CONTROL);
+        assert_eq!(parsed.modifiers, MOD_CONTROL as u32);
         assert_eq!(parsed.vk_code, 'A' as u32);
     }
